chore: rename to -ts, move unit tests and add operation assertion
