+++ nodejs22.x/hello-durable-ts/{{cookiecutter.project_name}}/hello-world/.npmignore
