+++ nodejs22.x/hello-durable-ts/{{cookiecutter.project_name}}/hello-world/jest.config.ts
@@ -1,7 +1,7 @@
 export default {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  testMatch: ['**/*.test.ts'],
+  testMatch: ['**/tests/unit/*.test.ts'],
   transform: {
     '^.+\\.ts$': 'ts-jest'
   },
