+++ nodejs24.x/hello-durable-ts/{{cookiecutter.project_name}}/hello-world/package.json
