+++ nodejs24.x/hello-durable-ts/{{cookiecutter.project_name}}/hello-world/tests/unit/app.test.ts
@@ -1,18 +1,13 @@
-import { lambdaHandler } from './app';
+import { lambdaHandler } from '../../app';
 import { APIGatewayProxyEvent } from 'aws-lambda';
-import { LocalDurableTestRunner } from '@aws/durable-execution-sdk-js-testing';
+import { LocalDurableTestRunner, OperationType, OperationStatus } from '@aws/durable-execution-sdk-js-testing';
 
 describe('Tests for durable hello world handler', () => {
     beforeAll(() => LocalDurableTestRunner.setupTestEnvironment({ skipTime: true }));
     afterAll(() => LocalDurableTestRunner.teardownTestEnvironment());
 
-    let runner: LocalDurableTestRunner<any>;
-
-    beforeEach(() => {
-        runner = new LocalDurableTestRunner({ handlerFunction: lambdaHandler });
-    });
-
     it('verifies successful response', async () => {
+        const runner = new LocalDurableTestRunner({ handlerFunction: lambdaHandler });
         const event: APIGatewayProxyEvent = {
             httpMethod: 'GET',
             path: '/hello',
@@ -22,9 +17,15 @@ describe('Tests for durable hello world handler', () => {
         const result = execution.getResult();
 
         expect(result.statusCode).toEqual(200);
-        expect(result.body).toBeDefined();
-        
         const body = JSON.parse(result.body);
         expect(body.message).toEqual('Hello, World!');
+
+        // Verify durable execution recorded one step operation
+        const operations = execution.getOperations();
+        expect(operations).toHaveLength(1);
+
+        const stepOperation = runner.getOperationByIndex(0);
+        expect(stepOperation.getType()).toBe(OperationType.STEP);
+        expect(stepOperation.getStatus()).toBe(OperationStatus.SUCCEEDED);
     });
 });
