+++ nodejs24.x/hello-durable-ts/{{cookiecutter.project_name}}/hello-world/tsconfig.json
@@ -12,5 +12,5 @@
     "resolveJsonModule": true,
     "moduleResolution": "node"
   },
-  "exclude": ["node_modules", "**/*.test.ts"]
+  "exclude": ["node_modules", "tests"]
 }
