+++ nodejs24.x/hello-durable-ts/{{cookiecutter.project_name}}/template.yaml
